+++ mcp_toolkit/devtools_mcp_extension/lib/core/devtools_core/service/vm_service_wrapper.dart
@@ -5,9 +5,6 @@
 // Code needs to match API from VmService.
 // ignore_for_file: avoid-dynamic
 
-/// @docImport '../shared/preferences/preferences.dart';
-library;
-
 import 'dart:async';
 import 'dart:convert';