+++ mcp_toolkit/mcp_toolkit/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.2
+
+- Added `kDefaultMaxErrors` and `maxErrors` constants to `ErrorMonitor` class to limit number of errors stored.
+- Added `kDebugMode` check to `MCPToolkitBinding.initialize` method.
+- Added `kDebugMode` check to `MCPToolkitExtensions.initializeServiceExtensions` method.
+- Added `kDebugMode` check to `MCPToolkitExtensions.registerServiceExtension` method to prevent adding entries in release mode.
+
 ## 0.1.1
 
 - Fixed documentation.